requirements.txt
pymupdf
openAI
frontend
python-dotenv